# docs/swagger.yaml
basePath: /
definitions:
  api.BulkScrapeRequest:
    properties:
      jobs:
        items:
          $ref: '#/definitions/api.ScrapeRequest'
        type: array
    required:
    - jobs
    type: object
  api.ScrapeRequest:
    properties:
      selectors:
        additionalProperties:
          type: string
        type: object
      strategy:
        example: custom
        type: string
      url:
        example: https://books.toscrape.com
        type: string
    required:
    - url
    type: object
  storage.DBJob:
    properties:
      created_at:
        type: string
      id:
        type: string
      result_title:
        type: string
      selectors:
        type: string
      status:
        type: string
      strategy:
        type: string
      updated_at:
        type: string
      url:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Distributed high-velocity Tor scraping API cluster
  title: GoScraper API
  version: "1.0"
paths:
  /db/reset:
    delete:
      description: Drops and recreates all tables
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Reset database
      tags:
      - system
  /health:
    get:
      description: Returns API health status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check
      tags:
      - system
  /health/deep:
    get:
      description: Checks DB connection, proxy pool, and worker queue depth
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Deep health check
      tags:
      - system
  /jobs:
    delete:
      description: Wipes all job records without dropping the table
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete all jobs
      tags:
      - jobs
    get:
      description: Returns paginated list of scrape jobs with optional status filter
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Page size (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: 'Filter: pending, processing, completed, failed'
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all jobs
      tags:
      - jobs
  /jobs/{id}:
    delete:
      description: Deletes a single job record by ID
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a job
      tags:
      - jobs
    get:
      description: Retrieves the details of a specific scrape job
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/storage.DBJob'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a scrape job
      tags:
      - jobs
  /jobs/{id}/retry:
    post:
      description: Requeues a failed job back into the worker pool
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Retry a failed job
      tags:
      - jobs
  /jobs/export:
    get:
      description: Returns all job records as a JSON array for download
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/storage.DBJob'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Export all jobs
      tags:
      - jobs
  /jobs/stats:
    get:
      description: Returns breakdown by strategy and success rates
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Job statistics
      tags:
      - jobs
  /metrics:
    get:
      description: Returns job counts and proxy pool summary
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: API metrics
      tags:
      - system
  /pool/node:
    get:
      description: Checks what exit IP a specific Tor node is currently using
      parameters:
      - description: Proxy URL e.g. http://tor-node-1:8118
        in: query
        name: url
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get current exit IP of a node
      tags:
      - pool
  /pool/reset:
    post:
      description: Reactivates all proxies and clears failure counters
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Reset proxy pool
      tags:
      - pool
  /pool/rotate:
    post:
      description: Sends NEWNYM signal to all active Tor nodes forcing fresh exit
        IPs
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Rotate all Tor circuits
      tags:
      - pool
  /pool/status:
    get:
      description: Returns health and stats for each proxy in the pool
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Proxy pool status
      tags:
      - pool
  /scrape:
    post:
      consumes:
      - application/json
      description: Submits a URL to the worker pool asynchronously
      parameters:
      - description: Scrape request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.ScrapeRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Submit a scrape job
      tags:
      - jobs
  /scrape/bulk:
    post:
      consumes:
      - application/json
      description: Submits multiple URLs in one request, each with its own strategy
      parameters:
      - description: Bulk scrape request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.BulkScrapeRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Submit multiple scrape jobs
      tags:
      - jobs
  /scrape/test:
    post:
      consumes:
      - application/json
      description: Scrapes a URL immediately and returns the result without saving
        to DB. Useful for testing selectors.
      parameters:
      - description: Scrape request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.ScrapeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Synchronous test scrape
      tags:
      - jobs
  /strategies:
    get:
      description: Returns all available scraping strategies and their descriptions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: List available strategies
      tags:
      - jobs
  /version:
    get:
      description: Returns Go version and build time
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Version info
      tags:
      - system
swagger: "2.0"